backend/requirements.cloud.txt
# 云存储可选依赖
# 根据需要安装相应的包

# AWS S3 存储
boto3>=1.34.0
botocore>=1.34.0

# Google Cloud Storage
google-cloud-storage>=2.10.0

# Azure Blob Storage
azure-storage-blob>=12.19.0

# 其他有用的包
aiofiles>=23.2.0  # 异步文件操作
python-multipart>=0.0.6  # 文件上传支持